feat:inserção de docker, e alteração das variaveis Long nas tabelas

+++ api/src/main/java/com/fawkes/api/Models/entity/OrderNote.java
@@ -3,6 +3,8 @@ package com.fawkes.api.Models.entity;
 import jakarta.persistence.*;
 import java.time.LocalDate;
 import lombok.*;
+import java.lang.*;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -12,7 +14,7 @@ public class OrderNote {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name = "noteNumber", unique = true, length = 45, columnDefinition = "CHAR(45)")
     private String numberNote;
